Split out teaching event building into separate controller
