+++ GetIntoTeachingApiTests/Controllers/TeachingEventsControllerTests.cs
@@ -191,23 +191,6 @@ namespace GetIntoTeachingApiTests.Controllers
             response.Should().BeOfType<UnauthorizedResult>();
         }
 
-        [Fact]
-        public void GetTeachingEventBuildings_ReturnsAllTeachingEventBuildings()
-        {
-            var mockBuildings = new List<TeachingEventBuilding>()
-            {
-                new TeachingEventBuilding() { AddressCity = "test" }
-            }
-            .AsQueryable();
-
-            _mockStore.Setup(mock => mock.GetTeachingEventBuildings()).Returns(mockBuildings);
-
-            var response = _controller.GetTeachingEventBuildings();
-
-            var ok = response.Should().BeOfType<OkObjectResult>().Subject;
-            ok.Value.Should().Be(mockBuildings);
-        }
-
         private static IEnumerable<TeachingEvent> MockEvents()
         {
             var event1 = new TeachingEvent() { Name = "Event 1", TypeId = 123 };
